se añade dotenvy para utilizar .env en el proyecto

+++ src/cors.rs
@@ -7,7 +7,7 @@ use rocket_cors::AllowedOrigins;
 
 use rocket_cors::Cors;
 
-pub fn run(port: &i32) -> Cors {
+pub fn run(port: &u16) -> Cors {
     let allowed_origins = AllowedOrigins::some_exact(&[
         format!("http://localhost:{port}"),
         format!("http://127.0.0.1:{port}"),
